tsweb, cmd/derper: move common web/debug stuff from derper to new tsweb
